students/jeongwook/requirements.txt
Django==4.0.4
django-cors-headers==3.11.0
mysqlclient==2.1.0
bcrypt==3.2.0
six==1.16.0
pycparser==2.21
cffi==1.15.0
pyjwt==2.3.0